modified: cleanup, changed naming, fixed sync issues
