--- src/actions/config.js
@@ -1,15 +1,0 @@
-const SET_SCREEN_HEIGHT = "SET_SCREEN_HEIGHT"
-
-
-function setScreenHeight(screenHeight) {
-    return ({
-        type: SET_SCREEN_HEIGHT,
-        screenHeight
-    })
-}
-
-
-export {
-    SET_SCREEN_HEIGHT,
-    setScreenHeight
-}
