+++ src/components/FilePicker/FilePicker.jsx
@@ -4,6 +4,7 @@ import { Button } from 'react-native-paper';
 import * as DocumentPicker from 'expo-document-picker';
 import * as FileSystem from 'expo-file-system';
 import Papa from 'papaparse';
+
 import { jsonToTemplate } from './jsonToTemplate'
 
 function FilePicker(props) {
@@ -13,10 +14,11 @@ function FilePicker(props) {
             if (type === "cancel") {
                 return
             }
-            const csv = await FileSystem.readAsStringAsync(uri)
 
+            const csv = await FileSystem.readAsStringAsync(uri)
             const trainingTemplate = jsonToTemplate(Papa.parse(csv).data)
             props.appendProgram(trainingTemplate)
+
         } catch (err) {
             console.error(err)
         }
