+++ src/components/Program/Exercise.jsx
@@ -1,5 +1,5 @@
-import React, { useState, useEffect } from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import React, { useState } from 'react';
+import { StyleSheet } from 'react-native';
 import { DataTable, List, Surface, Checkbox, TextInput } from 'react-native-paper';
 import { exerciseStatus } from '../../constants/contants'
 
@@ -10,7 +10,7 @@ function Exercise(props) {
     const [expanded, setExpanded] = useState(props.expanded);
     const [notes, setNotes] = useState('')
     return (
-        <Surface style={styles.item} >
+        <Surface style={styles.item}>
             <List.Accordion
                 title={exercise.name}
                 description={exerciseStatus[status].text}
